Renamed welcome.dart to splash.dart. Made views from a Center child to Expanded child.
